requirements.txt
# PyTorch GammaNet Requirements

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
scikit-image>=0.19.0

# Configuration and utilities  
pyyaml>=6.0
omegaconf>=2.3.0
tqdm>=4.62.0

# Training infrastructure
accelerate>=0.24.0
wandb>=0.15.0
tensorboard>=2.11.0

# Data processing
opencv-python>=4.5.0
pillow>=9.0.0
albumentations>=1.3.0

# Development tools
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.990

# Optional: For in silico experiments
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.4.0
